refactor app shell state hooks
